backend/requirements.txt
django>=2.0.4
django-auth-ldap>=1.4.0
djangorestframework>=3.8.1
djangorestframework-jwt>=1.11.0
djangorestframework-csv>=2.1.0
biopython>=1.71
uwsgi>=2.0.17
psycopg2-binary>=2.7.4
requests>=2.18.4
brotlipy>=0.7.0
whitenoise>=3.3.1
